Added Camera class, reorganized model drawing, and added spike model
Camera position information is modified based on mouse movement. VertexPositionTexture[] objects are created with their respective BasicEffects in the main file, and both are added to lists verticesObjects and basicEffectsForVerticesObjects. Models are read in from files and their effects are changed appropriately, then added to a list modelObjects. Then these lists, along with camera position information, are passed to the camera to be drawn.