Reorganized drawing and enabled lighting in models
This now allows drawing multiple instances of the same model and choosing in which order we draw the models. Instead of Camera.Update taking in lists of models and drawing them, it just sets the currentViewMatrix and currentProjectionMatrix. Then in the main file, we call Camera.DrawModel and Camera.DrawVertices within the Draw function as necessary. Also, enabled lighting in models (loaded from files, not arrays of vertices).